+++ docs_src/docs/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## V0.2.0
 
+**ADD**
+
+- [`monad_std.Option`](Api Document/option.md):
+    - [`__add__/+`](Api Document/option.md#monad_std.option.Option.__add__): Add alias for `bool_and`.
+- [`monad_std.Result`](Api Document/result.md):
+    - [`__add__/+`](Api Document/result.md#monad_std.result.Result.__add__): Add alias for `bool_and`.
+
 **FIX**
 
 - [`monad_std.Option`](Api Document/option.md):